moving AmbientLight to light/ambient_light.rs
